+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/sync/SyncTestExecutor.java
@@ -229,7 +229,7 @@ public class SyncTestExecutor {
             long listenerAncientThreshold = listenerExpiredThreshold;
             final double listenerDif = listenerMaxIndicator - listenerExpiredThreshold;
             if (listenerDif >= 3) {
-                listenerAncientThreshold += Math.floor(listenerDif / 3);
+                listenerAncientThreshold += (long) Math.floor(listenerDif / 3);
             } else if (listenerDif == 2) {
                 listenerAncientThreshold++;
             }
@@ -237,7 +237,7 @@ public class SyncTestExecutor {
             long callerAncientThreshold = callerExpiredThreshold;
             final double callerDif = callerMaxIndicator - callerExpiredThreshold;
             if (callerDif >= 3) {
-                callerAncientThreshold += Math.floor(callerDif / 3);
+                callerAncientThreshold += (long) Math.floor(callerDif / 3);
             } else if (callerDif == 2) {
                 callerAncientThreshold++;
             }
